@@ -31,7 +31,7 @@
     "watch": "npm-watch start"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=8.0.0"
   },
   "repository": {
     "type": "git",
@@ -56,16 +56,15 @@
     "@types/moment-timezone": "^0.5.6",
     "@types/node": "latest",
     "@types/node-fetch": "^2.1.1",
-    "@types/react": "^16.3.14",
-    "@types/react-dom": "^16.0.5",
+    "@types/react": "^16.4.14",
+    "@types/react-dom": "^16.0.7",
+    "@types/winston": "^2.4.4",
     "ava": "^0.23.0",
     "npm-watch": "^0.3.0",
     "rimraf": "^2.6.2",
-    "typescript": "latest"
+    "typescript": "^3.0.3"
   },
   "dependencies": {
-    "@ournet/places-domain": "^0.1.4",
-    "@ournet/weather-domain": "^0.1.4",
     "apollo-cache-inmemory": "^1.2.1",
     "apollo-client": "^2.3.1",
     "body-parser": "^1.18.3",
@@ -81,8 +80,8 @@
     "ms": "^2.1.1",
     "node-fetch": "^2.1.2",
     "ournet.links": "^0.4.2",
-    "ournet.logger": "^0.1.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
+    "react": "^16.5.0",
+    "react-dom": "^16.5.0",
+    "winston": "^3.1.0"
   }
 }
